Settings: Use uint getter/setter for more fields

+++ src/core/settings.h
@@ -93,8 +93,8 @@ struct Settings
 
   bool rewind_enable : 1 = false;
   float rewind_save_frequency = 10.0f;
-  u32 rewind_save_slots = 10;
-  u32 runahead_frames = 0;
+  u8 rewind_save_slots = 10;
+  u8 runahead_frames = 0;
 
   GPURenderer gpu_renderer = DEFAULT_GPU_RENDERER;
   std::string gpu_adapter;
